Schema refinements
Rename DatasetStatistic to DatasetAttibute, because we also want
to store non-numeric data and don't want the name to be misleading.
Also, introduce Attribute table, so that we can store descriptions
and labels without repeating ourselves.

Introduce a label column for datasets - the name column is intended
to be computer-readable, the label is human-readable and nicely
formatted. Analogous to (and often derived from) analysis.logic_name
and analysis_description.display_label from a core db.

Finally, increase the length of a couple of label fields.