Adding fixtures for ensembl_metadata models.
We have three model groupings, which reflects the three
different 'things' we care about: assembly, genome, and release.
So it makes sense to split our fixtures along these lines,
since we might want to test one 'thing' in isolation.
Furthermore, the assembly_sequence data can be quite large,
so we put that into its own file, to save having to load it
unless we really need it.