Test datasets (#3)
Test data added for a small set of species, covering a range of scenarios, including: multiple assemblies for one species, multiple genesets for one assembly, a strain group, and species in multiple sites/releases.

We have three model groupings, which reflects the three different 'things' we care about: assembly, genome, and release. So it makes sense to split our fixtures along these lines, since we might want to test one 'thing' in isolation. Furthermore, the assembly_sequence data can be quite large, so we put that into its own file, to save having to load it unless we really need it.

Taxonomy tests and fixtures updated to be in sync with the metadata fixtures.